Bug #1164785 use Andoird standard TAG for ISO
[root cause  ] code optimise
[changes     ] code optimise
[side effects] no
[self test   ]: yes
[download normally]: yes
[power on/off normally]: yes
[do common repository/branch inspection]: yes
[is there dependence]: no
[confirm dependent commit]: no
[board]: sharkl5Pro
[test case]: manual iso test
[reviewers   ] victor
[change_type ] feature_bugfix
[tag_product ] common

+++ hal3_2v6/SprdCamera3Setting.h
@@ -629,7 +629,7 @@ class SprdCamera3Setting {
                                        int8_t *convertDrvMode);
     int androidSceneModeToDrvMode(uint8_t androidScreneMode,
                                   int8_t *convertDrvMode);
-    int androidIsoToDrvMode(uint8_t androidIso, int8_t *convertDrvMode);
+    int androidIsoToDrvMode(int32_t androidIso, int8_t *convertDrvMode);
     int androidEffectModeToDrvMode(uint8_t androidEffectMode,
                                    int8_t *convertDrvMode);
     int androidAwbModeToDrvAwbMode(uint8_t androidAwbMode,
